src/main/kotlin/com/example/springbootkotlinhexagonaldemo/application/service/UpdateUserByIdService.kt
package com.example.springbootkotlinhexagonaldemo.application.service

import com.example.springbootkotlinhexagonaldemo.application.port.persistence.ReadUserPort
import com.example.springbootkotlinhexagonaldemo.application.port.persistence.WriteUserPort
import com.example.springbootkotlinhexagonaldemo.application.usecase.UpdateUserByIdUseCase
import com.example.springbootkotlinhexagonaldemo.domain.entity.User
import com.example.springbootkotlinhexagonaldemo.domain.model.UserModification
import com.example.springbootkotlinhexagonaldemo.domain.type.id.UserId
import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional

@Transactional
@Service
class UpdateUserByIdService(
    private val readUserPort: ReadUserPort,
    private val writeUserPort: WriteUserPort,
) : UpdateUserByIdUseCase {

    override fun updateUserById(id: UserId, userModification: UserModification): User {
        val user = readUserPort.findById(id)
        requireNotNull(user)

        user.update(userModification)
        writeUserPort.update(user)
        return user
    }
}